chore(cleanup): reorganize directory structure and add homepage module
